refactor: check nickname component를 text input 컴포넌트와 병합
